progress and separation of responsibility
