+++ app/routes/funding/what-funding.js
@@ -1,4 +1,3 @@
-const { required } = require('joi')
 const Joi = require('joi')
 const cache = require('../../cache')
 const getFunding = require('../../funding/get-funding')
@@ -9,11 +8,15 @@ module.exports = [{
   path: '/what-funding',
   options: {
     handler: async (request, h) => {
-      const cacheMock = require('../../cache_mock')
-      cache.set(request, cacheMock[0]) // mimics the cache being built up before this page
+      if (!('agreement' in await cache.get(request))) {
+        console.log('mocking cache')
+        const cacheMock = require('../../cache_mock')
+        cache.set(request, cacheMock[0]) // mimics the cache being built up before this page
+      }
 
       const { agreement } = await cache.get(request)
-      const { funding } = agreement
+      const funding = 'funding' in agreement ? agreement.funding : []
+      // const funding = agreement?.funding ?? []
 
       const eligibleFunding = await getFunding(request)
 
@@ -45,38 +48,24 @@ module.exports = [{
       }
     },
     handler: async (request, h) => {
-      // const { standard } = request.payload
-      // const { data, agreement } = await cache.get(request)
+      const standard = typeof (request.payload) === 'string' ? [request.payload] : request.payload
+      const { data, agreement } = await cache.get(request)
 
-      // const funding = data?.eligibleFunding.filter(x => standard.includes(x.code)) ?? []
+      const funding = data?.eligibleFunding.filter(x => standard.includes(x.code)) ?? []
 
-      // if (!funding.length) {
-      //   return h.redirect('/what-funding')
-      // }
-
-      // for (const option in agreement.action) {
-      //   agreement.action[option].active = funding.some(x => x.code === option)
-      // }
-      // agreement.funding = funding.map(x => x.code)
-
-      // await cache.update(request, { agreement })
-      // return h.redirect('/how-much')
-
-      const { agreement } = await cache.get(request)
-      const { standard } = request.payload
-
-      if (typeof (standard) === 'string') {
-        agreement.action[standard] = { actionsComplete: true }
-      } else {
-        standard.map(option => {
-          agreement.action[option] = { actionsComplete: true }
-          return 1
-        })
+      if (!funding.length) {
+        return h.redirect('/what-funding')
       }
+
+      for (const option in agreement.action) {
+        agreement.action[option].active = funding.some(x => x.code === option)
+        agreement.action[option].actionsComplete = true
+      }
+      agreement.funding = funding.map(x => x.code)
 
       cache.update(request, { agreement })
 
-      return h.redirect('/task-list')
+      return h.redirect('/how-much')
     }
   }
 }]
